tiered commands and new statreset command
